Turn the fan on when the outside and inside temps are the same
Since the weather conditions API seems to lag a little behind
the actual outside temperature, turning on the fan when the
outside and inside temps are the same is likely to bring in
cooler air. It will at least get the air moving.

Also fixed a compile issue introduced during the cleanup of the
script.